Move player-related stuff to Pawn folder/namespace
