+++ code/Pawn/HoldPose.cs
@@ -1,0 +1,9 @@
+﻿namespace TerryForm.Pawn
+{
+	public enum HoldPose
+	{
+		None,
+		Bazooka,
+		Shotgun
+	}
+}
