+++ code/States/WaitingState.cs
@@ -1,4 +1,4 @@
-﻿
+﻿using Sandbox;
 
 namespace TerryForm.States
 {
